feat(config): add nuxt environment browser-testing support
