+++ src/runtime/entry.ts
@@ -1,19 +1,12 @@
 import { vi } from 'vitest'
+import { setupNuxt } from './shared/nuxt'
 
 if (
   typeof window !== 'undefined'
   // @ts-expect-error undefined property
   && window.__NUXT_VITEST_ENVIRONMENT__
 ) {
-  const { useRouter } = await import('#app/composables/router')
-  // @ts-expect-error alias to allow us to transform the entrypoint
-  await import('#app/nuxt-vitest-app-entry').then(r => r.default())
-  // We must manually call `page:finish` to snc route after navigation
-  // as there is no `<NuxtPage>` instantiated by default.
-  const nuxtApp = useNuxtApp()
-  await nuxtApp.callHook('page:finish')
-  useRouter().afterEach(() => nuxtApp.callHook('page:finish'))
-
+  await setupNuxt()
   vi.resetModules()
 }
